Vast redesign of the the graph file.
Nodes are now simple strings, except for two special cases: anonymous
nodes, which act like blanks, and type nodes, which act like "type" but
do not equal each other.

A map contains a set of nodes, and a set of edges. All links are in the
edges.

To link to nodes, call link and unlink on their map.

Read the description of the add function for the new modifications to
the design